keynote/requirements.txt
streamlit
extra-streamlit-components
google-cloud-storage
google-cloud-aiplatform
pytubefix